Remove dev_go
Go and specifically the GOPATH do not play nice with containers. You
would have to run this container at the top of your GOPATH every time,
and cd to the project you want, which is trash.

If you want to write Go, use a big-ass VM or just install Go locally.
It's stupidly un-containerable.