Add init script to support different UIDs
So this was a fun suprise I didn't forsee: if your UID isn't 1000, you
can't mount things because the UIDs won't match between the `developer`
account and the host user.

This was brought to light when I switched my laptop to use a FreeIPA
Enterprise Login, which then had my UID as some ridiculously large UID.

Luckily, this was a pretty quick fix so this should just be a drop-in to
the other containers to make things work.